# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TrustLoopGuard API
  description: Real-time guardrail runtime for AI agents.
  license:
    name: Apache-2.0
  version: 0.0.1
paths:
...
  /v1/financial/actions:
    get:
      tags:
      - financial
      operationId: list_actions
      responses:
        '200':
          description: Financial actions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialActionListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - financial
      operationId: create_action
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFinancialActionRequest'
        required: true
      responses:
        '201':
          description: Financial action created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialActionRecord'
        '400':
          description: Malformed or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/actions/{id}:
    get:
      tags:
      - financial
      operationId: get_action
      parameters:
      - name: id
        in: path
        description: Financial action id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Financial action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialActionRecord'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Financial action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/actions/{id}/approve:
    post:
      tags:
      - financial
      operationId: approve_action
      parameters:
      - name: id
        in: path
        description: Financial action id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Financial action authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialActionRecord'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Financial action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/actions/{id}/deny:
    post:
      tags:
      - financial
      operationId: deny_action
      parameters:
      - name: id
        in: path
        description: Financial action id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Financial action denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialActionRecord'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Financial action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/actions/{id}/execute:
    post:
      tags:
      - financial
      operationId: execute_action
      parameters:
      - name: id
        in: path
        description: Financial action id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Financial action executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialActionRecord'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Financial action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/actions/{id}/outcomes:
    get:
      tags:
      - financial
      operationId: list_action_outcomes
      parameters:
      - name: id
        in: path
        description: Financial action id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Financial action outcomes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialOutcomeListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Financial action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - financial
      operationId: record_action_outcome
      parameters:
      - name: id
        in: path
        description: Financial action id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinancialActionOutcome'
        required: true
      responses:
        '201':
          description: Financial action outcome recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialActionOutcome'
        '400':
          description: Malformed or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Financial action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/approval-requests:
    get:
      tags:
      - financial
      operationId: list_approval_requests
      responses:
        '200':
          description: Financial approval requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialApprovalRequestListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/mandates:
    get:
      tags:
      - financial
      operationId: list_mandates
      responses:
        '200':
          description: Financial mandates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialMandateListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - financial
      operationId: create_mandate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFinancialMandateRequest'
        required: true
      responses:
        '201':
          description: Financial mandate created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialMandate'
        '400':
          description: Malformed or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/mandates/{id}/revoke:
    post:
      tags:
      - financial
      operationId: revoke_mandate
      parameters:
      - name: id
        in: path
        description: Financial mandate id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Financial mandate revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialMandate'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Financial mandate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/policies:
    get:
      tags:
      - financial
      operationId: list_policies
      responses:
        '200':
          description: Financial spending controls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialPolicyListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - financial
      operationId: create_policy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFinancialPolicyRequest'
        required: true
      responses:
        '201':
          description: Financial spending control created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialPolicyRecord'
        '400':
          description: Malformed or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/financial/receipts/{id}:
    get:
      tags:
      - financial
      operationId: get_receipt
      parameters:
      - name: id
        in: path
        description: Financial receipt id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Financial receipt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialReceipt'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Financial receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/policies:
    get:
      tags:
      - policies
      summary: '`GET /v1/policies` — list active policy summaries.'
      operationId: list_policies
      responses:
        '200':
          description: All active policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - policies
      summary: '`POST /v1/policies` — create or update a policy from YAML or JSON.'
      operationId: upsert_policy
      requestBody:
        description: Policy document, YAML or JSON
        content:
          application/yaml:
            schema:
              type: string
        required: true
      responses:
        '201':
          description: Policy created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDocument'
        '400':
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Policy failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    ApiError:
      type: object
      description: |-
        Canonical error envelope returned on non-2xx responses from every
        TrustLoopGuard endpoint. SDKs deserialize this body to produce typed
        errors; integrators don't have to inspect status codes by hand.

        The shape is intentionally minimal — `code` drives SDK-side fan-out,
        `message` is for logs, `retriable` tells callers whether the same
        request may be retried, and `details` is opaque so the server can
        add validation field paths without breaking SDKs.
      required:
      - code
      - message
      - retriable
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        details:
          description: |-
            Opaque structured details (e.g. validation field path).
            Defaults to `null`; servers may add fields without breaking SDKs.
        message:
          type: string
        retriable:
          type: boolean
          description: |-
            Whether the caller may retry the same request without modification.
            SDKs honor `Retry-After` when present in addition to this flag.
    ApiErrorCode:
      type: string
      description: |-
        Stable error code dictionary. Add variants here when introducing new
        failure modes; never repurpose an existing variant — SDK callers may
        be branching on it.
      enum:
      - invalid
      - unauthorized
      - forbidden
      - not_found
      - gone
      - unprocessable
      - rate_limited
      - internal
      - unavailable
...
    CreateFinancialActionRequest:
      type: object
      required:
      - idempotency_key
      - action
      properties:
        action:
          $ref: '#/components/schemas/FinancialAction'
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceRef'
        execute:
          type: boolean
        idempotency_key:
          type: string
    CreateFinancialMandateRequest:
      type: object
      required:
      - principal_id
      properties:
        expires_at:
          type:
          - string
          - 'null'
        id:
          type:
          - string
          - 'null'
        metadata: {}
        principal_id:
          type: string
        scope: {}
        starts_at:
          type:
          - string
          - 'null'
        version:
          type:
          - integer
          - 'null'
          format: int32
...
    EvidenceRef:
      type: object
      required:
      - source
      - source_id
      - kind
      properties:
        kind:
          type: string
        metadata: {}
        observed_at:
          type:
          - string
          - 'null'
        source:
          type: string
        source_id:
          type: string
...
    FinancialActionListResponse:
      type: object
      required:
      - actions
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/FinancialActionRecord'
    FinancialActionOutcome:
      type: object
      required:
      - action_id
      - status
      - reversal_capability
      - recovery_status
      - occurred_at
      properties:
        action_id:
          type: string
        final_loss_amount:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/MoneyAmount'
        metadata: {}
        occurred_at:
          type: string
        provider_reference:
          type:
          - string
          - 'null'
        provider_status:
          type:
          - string
          - 'null'
        recovery_status:
          $ref: '#/components/schemas/RecoveryStatus'
        reversal_capability:
          $ref: '#/components/schemas/ReversalCapability'
        status:
          $ref: '#/components/schemas/FinancialActionOutcomeStatus'
...
    FinancialActionRecord:
      type: object
      required:
      - id
      - workspace_id
      - status
      - action
      - created_at
      - updated_at
      properties:
        action:
          $ref: '#/components/schemas/FinancialAction'
        created_at:
          type: string
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceRef'
        id:
          type: string
        status:
          $ref: '#/components/schemas/FinancialActionStatus'
        status_reason:
          type:
          - string
          - 'null'
        updated_at:
          type: string
        workspace_id:
          type: string
...
    FinancialApprovalRequestListResponse:
      type: object
      required:
      - approval_requests
      properties:
        approval_requests:
          type: array
          items:
            $ref: '#/components/schemas/FinancialApprovalRequest'
...
    FinancialMandate:
      type: object
      required:
      - id
      - workspace_id
      - version
      - status
      - principal_id
      - created_at
      - updated_at
      properties:
        created_at:
          type: string
        expires_at:
          type:
          - string
          - 'null'
        id:
          type: string
        metadata: {}
        principal_id:
          type: string
        scope: {}
        starts_at:
          type:
          - string
          - 'null'
        status:
          $ref: '#/components/schemas/FinancialMandateStatus'
        updated_at:
          type: string
        version:
          type: integer
          format: int32
        workspace_id:
          type: string
    FinancialMandateListResponse:
      type: object
      required:
      - mandates
      properties:
        mandates:
          type: array
          items:
            $ref: '#/components/schemas/FinancialMandate'
...
    FinancialOutcomeListResponse:
      type: object
      required:
      - outcomes
      properties:
        outcomes:
          type: array
          items:
            $ref: '#/components/schemas/FinancialActionOutcome'
    FinancialPolicyListResponse:
      type: object
      required:
      - policies
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/FinancialPolicyRecord'
...
    FinancialReceipt:
      type: object
      required:
      - id
      - action_id
      - created_at
      properties:
        action_id:
          type: string
        created_at:
          type: string
        id:
          type: string
        ledger_event_ids:
          type: array
          items:
            type: string
        proof: {}
        trace_id:
          type:
          - string
          - 'null'
...
    PolicyDocument:
      type: object
      required:
      - id
      - severity
      - enabled
      - source_yaml
      properties:
        description:
          type:
          - string
          - 'null'
        enabled:
          type: boolean
        family:
          $ref: '#/components/schemas/PolicyFamily'
        id:
          type: string
        severity:
          $ref: '#/components/schemas/Severity'
        source_yaml:
          type: string
...
    PolicyListResponse:
      type: object
      required:
      - policies
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/PolicySummary'
...